# backend/cube/docs/swagger.yaml
basePath: /api
definitions:
...
  UserStatInfoResponse:
    properties:
      edit:
        items:
          $ref: '#/definitions/usersstat.StatInfo'
        type: array
      registration:
        items:
          $ref: '#/definitions/usersstat.StatInfo'
        type: array
    type: object
...
  offset.Info:
    properties:
      id:
        type: integer
      neuron_id:
        type: integer
      offset:
        type: number
    type: object
...
  userblockUserResponse:
    properties:
      blocked_until:
        example: "2025-08-09T15:00:00.053Z"
        type: string
      id:
        example: f6457bdf-4e67-4f05-9108-1cbc0fec9405
        type: string
    type: object
  usersstat.StatInfo:
    properties:
      id:
        example: 3f8bf2a3-01cc-4c4a-9759-86cec9cf8da9
        type: string
      time:
        example: 2006-01-02T15:04:05Z07:00
        type: string
    type: object
  weight.Info:
    properties:
      id:
        type: integer
      link_id:
        type: integer
      weight:
        type: number
    type: object
  weights.Info:
    properties:
      id:
        type: string
      name:
        type: string
      offsets:
        items:
          $ref: '#/definitions/offset.Info'
        type: array
      weights:
        items:
          $ref: '#/definitions/weight.Info'
        type: array
    type: object
...
host: localhost
info:
  contact: {}
  description: This is cube server.
  license:
    name: MIT
    url: https://mit-license.org/
  title: Cube API
  version: "1.0"
paths:
  /v1/blocks/users/{user_id}:
    delete:
      description: Deletes user block info by user id
      parameters:
      - description: User ID to unblock
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: User unblocked
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Failed to delete user block info from storage
      security:
      - ApiKeyAuth: []
      summary: Delete user block info
      tags:
      - blocks
    get:
      description: Find such users info as id and block time
      parameters:
      - description: UserId to search for
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users block info found
          schema:
            $ref: '#/definitions/userblockUserResponse'
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Failed to get user info from storage
      security:
      - ApiKeyAuth: []
      summary: Find user block info
      tags:
      - blocks
    patch:
      description: Blocks user until specified moment
      parameters:
      - description: User ID to block
        in: path
        name: user_id
        required: true
        type: string
      - description: Time to block until
        in: query
        name: until
        required: true
        type: string
      responses:
        "200":
          description: User blocked
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: 'Failed to block user '
      security:
      - ApiKeyAuth: []
      summary: Block user
      tags:
      - blocks
...
  /v1/users/stats:
    get:
      description: Get such user stat info as registration and edit stat per period
      parameters:
      - description: Time to start from, RFC3339
        format: '"2006-01-02T15:04:05Z07:00"'
        in: query
        name: from
        type: string
      - description: Time to stop at, RFC3339
        format: '"2006-01-02T15:04:05Z07:00"'
        in: query
        name: to
        type: string
      - description: Search for load stat
        in: query
        name: load
        type: boolean
      - description: Search for update stats
        in: query
        name: update
        type: boolean
      responses:
        "200":
          description: Users stat info found
          schema:
            items:
              $ref: '#/definitions/UserStatInfoResponse'
            type: array
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "500":
          description: Failed to get user stat info
      security:
      - ApiKeyAuth: []
      summary: Get users stat info
      tags:
      - users
  /v1/weights:
    get:
      description: Find such model info as id, username, email and fullname
      parameters:
      - description: Structure ID to search for
        in: query
        name: structure_id
        required: true
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Page size for pagination
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: Model weights info found
          schema:
            items:
              $ref: '#/definitions/weights.Info'
            type: array
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Failed to get model weights info from storage
      security:
      - ApiKeyAuth: []
      summary: Find structure weights info
      tags:
      - weights
    post:
      consumes:
      - multipart/form-data
      description: Adds model weights info to existing model
      parameters:
      - description: Model ID to add weights to
        in: formData
        name: model_id
        required: true
        type: string
      - description: Model Weights title to add
        in: formData
        name: weights_title
        required: true
        type: string
      - description: Model Weights to add
        in: formData
        name: weights
        required: true
        type: file
      responses:
        "200":
          description: Weights added
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Invalid token, user id not specified
        "500":
          description: Failed to create model weights info
      security:
      - ApiKeyAuth: []
      summary: Create new model weights info
      tags:
      - weights
  /v1/weights/{weight_id}:
    delete:
      description: Deletes model info from any user
      parameters:
      - description: Model ID to delete
        in: path
        name: weight_id
        required: true
        type: string
      responses:
        "200":
          description: Model info deleted
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Invalid token, user id not specified
        "404":
          description: Not Found
        "500":
          description: Failed to delete model info from storage
      security:
      - ApiKeyAuth: []
      summary: Delete model info
      tags:
      - weights
    get:
      description: Find such model weights info as id, name, link weights and neuron
        offsets
      parameters:
      - description: Weight ID to search for
        in: path
        name: weight_id
        required: true
        type: string
      responses:
        "200":
          description: Model weights info found
          schema:
            items:
              $ref: '#/definitions/weights.Info'
            type: array
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Failed to get model weights info from storage
      security:
      - ApiKeyAuth: []
      summary: Find model weights info
      tags:
      - weights
    patch:
      consumes:
      - multipart/form-data
      description: Update such model info as weights, weights titles
      parameters:
      - description: Model Weights ID to update
        in: path
        name: weight_id
        required: true
        type: string
      - description: Model Weights Title to set
        in: formData
        name: weight_title
        type: string
      - description: Model Weights to Update/Add
        in: formData
        name: weights
        type: file
      responses:
        "200":
          description: Model info updated
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
        "403":
          description: Invalid token, user id not specified
        "404":
          description: Not Found
        "500":
          description: Failed to update model info
      security:
      - ApiKeyAuth: []
      summary: Update model info
      tags:
      - weights
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"